Add files via upload
Concrete is most important man-made material for building. Concrete is
made up of three basic components water, aggregate, and portland
cement. Quality is the most important property for material used in
building, if the quality of concrete is less, then the build can’t be stable but
if we use the best quality concrete then the building is stable, safety and
longevity of structure is maintained. To know that this concrete is quality
proof or not the Compressive Strength of Concrete is used to determine the
quality of Concrete, the quality is checked by using standard crushing test
on a concrete cylinder. Concrete strength is also considered as a key factor
in obtaining the desired durability. For testing strength, it will take 7 or 14
or 28 days this is a long time. So by using machine learning and auto
machine learning techniques with Keras Sequential model and Auto
keras this problem can be solved, it predicts the strength of the concrete. It
reduces lot of efforts and also it will predict that in how much quantity we
have to use, which raw material is used for good compressive strength.
Through rigorous evaluation and comparison, we demonstrate the
superiority of our model over traditional methods. This study not only
contributes to improved engineering practices but also underscores the
potential of machine learning in optimizing material design and
construction processes.